+++ Source/AllOverIt.Csv/Exceptions/CsvSerializerException.cs
@@ -3,16 +3,16 @@
 namespace AllOverIt.Csv.Exceptions
 {
     /// <summary>Represents an error that occurs during the export of data to CSV.</summary>
-    public sealed class CsvExportException : Exception
+    public sealed class CsvSerializerException : Exception
     {
         /// <summary>Default constructor.</summary>
-        public CsvExportException()
+        public CsvSerializerException()
         {
         }
 
         /// <summary>Constructor.</summary>
         /// <param name="message">The exception message.</param>
-        public CsvExportException(string message)
+        public CsvSerializerException(string message)
             : base(message)
         {
         }
@@ -20,7 +20,7 @@ namespace AllOverIt.Csv.Exceptions
         /// <summary>Constructor.</summary>
         /// <param name="message">The exception message.</param>
         /// <param name="innerException">The inner exception.</param>
-        public CsvExportException(string message, Exception innerException)
+        public CsvSerializerException(string message, Exception innerException)
             : base(message, innerException)
         {
         }